+++ lib/src/builders/handlers/messages/when_result_descriptor.dart
@@ -28,6 +28,7 @@ MessageHandlerDescriptor<C> makeWhenResultMessageDescriptor<M, D, C, T>(
     conditions,
     messageName,
     label,
+    {},
   );
 
   return MessageHandlerDescriptor<C>(